+++ NEWS.md
@@ -2,7 +2,7 @@
 ### Enhancements
 * Refactored the internal code of all the modules to optimize their performance.
 * Refactored `tm_g_scatterplot` to compute trend line statistics using `ggpmisc::stat_poly_eq` to simplify code and to have labels for each group on the plot.
-* Refactored `tm_g_scatterplot` to enable filtering via the Encoding Panel using the new `filter_spec` functionalities.
+* Refactored `tm_g_scatterplot` to enable filtering via the Encoding Panel using the new `filter_spec` functionality.
 * Added support for logical variables in `tm_variable_browser`. 
 * Updated `tm_outliers` to handle non-CDISC datasets.
 * Updated UI of `tm_variable_browser` for factor and character variables with more than 30 levels.